Remove `.ruby-gemset` from the project
Use of gemsets is only really needed by [rvm](https://rvm.io/) users, and use of rvm is a choice of the developer, not a requirement of the project.

Hence this change removes the file from the project, and also adds a `.gitignore` entry so those who wish to use gemsets can do so without issue.