Switch DEFRA to Defra and expand desc. in gemspec
Even though Defra is actually an acronym for Department for Environment, Food & Rural Affairs and would be expected to appear as DEFRA, in our comms we actually use 'Defra'. See our main page on [GOV.uk](https://www.gov.uk/government/organisations/department-for-environment-food-rural-affairs) for examples of this.

Also expanded the descriptions and summary to make it clear this is ruby based.