Update gemspec to match change in project name
We have updated the project name from **Defra style** to **Defra ruby style** to ensure this is not confused as a gem used to govern all styles in Defra.

As such this meant we needed to change the name of the gemspec file, plus some of the details in it.

This also fixes #2