Remove frozen string literal exclusion for rspec
The exclusion was added for fear that specs would be affected by using frozen strings, however on reflection it was noted that we as yet have not experienced any issues where the comment is added.

Also the rspec team appear to be working on ensuring compliance with it (see <https://github.com/rspec/rspec-core/issues/2264>). So we've decided we should be adding this magic comment and exposing any issues upfront, rather than being caught by surprise later.