Add travis ci integration to the project
The project has one dependency, rubocop and no actual code files in this project. We have the project hooked up to [Dependabot](https://dependabot.com/) which is likely to update this dependency on occasion.

We want some form of CI to run when this happens to ensure that rubocop will still run using the default config.

We also want to know that as we change the default rules in `default.yml` rubocop will still run.

Hence this change integrates the project with Travis, and when travis runs the test is to see if rubocop runs.