Update project to enable auto deployments
As a shared gem that multiple services are dependent on, our pattern is to version and deploy the gem to rubygems.

This makes the changes more visible, and allows us to hook up Dependabot in such a way that when a change is made to this gem, it will automate updating the projects that use it and testing whether anything breaks.