Remove exclusion for config folder in projects
We spotted in a project recently that new initializers we were adding to the Rails `config/initializers` folder were not being picked up by rubocop.

A quick check of the default rules threw up that we had excluded the folder. The reason given is valid, but on reflection we make so may changes in this folder, and often add many new files that it makes sense to get everything consistent.

The downside is when starting a new project we may be forced to reformat the stuff Rails generates, but overall we think its better to have everything consistent, and that this isn't that onerous a task to do.