Fix Naming/FileName cop entry in default.yml
Spotted when using the latest version of defra ruby style in other repos that we were getting this error message.

```
Warning: unrecognized cop Naming/Filename found in .rubocop.yml
```

A quick google and spotted that the cop name is actually `Naming::FileName` so the entry should be `Naming/FileName` hence the warning message.

This change fixes the issue.