Set memoized instance variable name to an optional enforced leading underscore
https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Naming/MemoizedInstanceVariableName

When using an instance variable to cache an operation that we only want to do once, we
agree to use the format `@_method_name ||= SomeOperation`. This makes it clearer that
the instance variable should never be referred to directly but always throught its attr accessor
method.
Memoized variables are commonly used to allow a Lazy load / computation of the attribute itself.