Specify pessimistic version for Rubocop
Hakiri is flagging a security issue with the project

> [CVE-2017-8418](https://hakiri.io/technologies/rubocop/issues/6e12d31ab4de2e) - RuboCop 0.48.1 and earlier does not use /tmp in safe way, allowing local users to exploit this to tamper with cache files belonging to other users.

It has been patched in later versions, and that's what we are using throughout our projects. However because we don't specify a version in our gemspec Hakiri will keep flagging us, because theoretically a user could use our gem with an old version of rubocop.

So this update adds pessimistic versioning for the rubocop dependency. This should have the effect of saying this is the minimum version of rubocop that can be used, whilst still allowing us to take an future releases.