Replace logo with title
One day we might go back to using an icon as it helps make the project stand out. But if we do ideally the logo would have a consistent design with other `defra_ruby`, and all would have one.

So in the meantime to keep things consistent and avoid any questions, this change removes the existing name and logo image and uses a standard `#` level markdown title.