Update rubocop and enable all new cops
https://docs.rubocop.org/en/latest/versioning/#pending-cops

From version 0.80 of rubocop it appears that new cops when added will no longer be automatically default to enabled. Instead they will be `pending` and a warning message will be output advising users that they need to decide whether to enable the cop or not.

Though new cops do generally break our builds (!) we want to always be using the latest conventions and will deal with build issues as they arise. We also don't want to have to keep dipping into the config everytime a new cop is added.

So this change does 2 things

- it updates rubocop to the latest which includes the new pending cops strategy
- it adds a flag to our default config to always enable new cops